Prefill Prior Forecasts
The bug seemed to be in the confusion that scenario results held forecast values, but that stores an array of the averages after the scenario has concluded.

The fix was to fetch the scenario forecast for the current user and call the `handleForecastChange()` onload so the total is calculated if the forecast is prepoplated (small side effect is that on a new Scenario it displays 0).